Adjusting code to header renames on include/drivers
