+++ include/drivers/optimize_driver.h
@@ -1,5 +1,5 @@
 /*PGR-GNU*****************************************************************
-File: pickDeliver_driver.h
+File: optimize_driver.h
 
 Copyright (c) 2015 pgRouting developers
 Mail: project@pgrouting.org
