+++ include/drivers/vroom_driver.h
@@ -26,8 +26,8 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
  ********************************************************************PGR-GNU*/
 
-#ifndef INCLUDE_DRIVERS_VROOM_VROOM_DRIVER_H_
-#define INCLUDE_DRIVERS_VROOM_VROOM_DRIVER_H_
+#ifndef INCLUDE_DRIVERS_VROOM_DRIVER_H_
+#define INCLUDE_DRIVERS_VROOM_DRIVER_H_
 #pragma once
 
 #include "c_types/typedefs.h"
@@ -61,4 +61,4 @@ extern "C" {
 }
 #endif
 
-#endif  // INCLUDE_DRIVERS_VROOM_VROOM_DRIVER_H_
+#endif  // INCLUDE_DRIVERS_VROOM_DRIVER_H_
