+++ src/pgr_pickDeliver/pickDeliverEuclidean.c
@@ -36,7 +36,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include "c_types/pickDeliveryOrders_t.h"
 #include "c_types/solution_rt.h"
 
-#include "drivers/pgr_pickDeliver/pickDeliverEuclidean_driver.h"
+#include "drivers/pgr_pickDeliverEuclidean_driver.h"
 
 PGDLLEXPORT Datum _vrp_pgr_pickdelivereuclidean(PG_FUNCTION_ARGS);
 PG_FUNCTION_INFO_V1(_vrp_pgr_pickdelivereuclidean);
