+++ src/pgr_pickDeliver/pickDeliverEuclidean_driver.cpp
@@ -26,7 +26,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  ********************************************************************PGR-GNU*/
 
 
-#include "drivers/pgr_pickDeliver/pickDeliverEuclidean_driver.h"
+#include "drivers/pgr_pickDeliverEuclidean_driver.h"
 
 #include <sstream>
 #include <utility>
